Updates `ApiClient` to use constants
As I'm building this client out, I've only been testing this against the
public api.envato.com however internally we have other environments (and
hostnames) that we can use for testing.

If I need to test this on another hostname, it's a right royal pain to
update all the references and then do the work. Instead, I'm defining a
`base_uri` in the Guzzle options which will allow me to set a default
hostname but pass in the option when I need to override it.

```php
$client = ApiClient::factory(array(
  'token' => 'xxx',
  'base_uri' => 'someotherhost.example.com'
));
```

This commit also pulls out the user agent and version into constants too
since we don't really need those to be changing but should be updated in
one spot if/when we do decide to.