Create a `RateLimiter` for managing rate limited responses
This introduces a new middleware class that is responsible for
gracefully handling when the server presents a HTTP 429 (Too many
requests). This HTTP status is used when a client has reached the
threshold of what has been deemed a suitable threshold by the service
owner and the client should slow their request rate or back off with the
requests.

There are a number of reasons why we would want a piece of middleware
like this.

1. Know when a request is getting rate limited and present that error
   code in a more useful manner that can be identified and tracked.
2. Don't just keep trying the same request over and over again. In some
   of the cases, continuing to request the resource whilst being rate
   limited results in your retry timestamp getting pushed out using
   expotential backoff. This is factored into the class by short
   circuiting the request and not actually sending it if the throttler
   is enabled.